Fix build

+++ yt/ytlib/table_chunk_format/column_writer_detail.cpp
@@ -57,7 +57,7 @@ void TColumnWriterBase::DumpSegment(TSegmentInfo* segmentInfo)
     BlockWriter_->WriteSegment(MakeRange(segmentInfo->Data));
 }
 
-i32 TColumnWriterBase::GetMetaSize() const
+i64 TColumnWriterBase::GetMetaSize() const
 {
     return MetaSize_;
 }
